+++ src/main/java/chess/domain/piece/strategy/KnightLegalMoveCheckStrategy.java
@@ -1,22 +1,22 @@
-package chess.domain.strategy;
+package chess.domain.piece.strategy;
 
 import chess.domain.board.Board;
 import chess.domain.square.Square;
 import chess.domain.square.dto.SquareDifferent;
 
-public class KingLegalMoveCheckStrategy implements LegalMoveCheckStrategy {
+public class KnightLegalMoveCheckStrategy implements LegalMoveCheckStrategy {
 
     @Override
     public boolean check(Square source, Square destination, Board board) {
         SquareDifferent diff = source.calculateDiff(destination);
 
-        int fileDiff = Math.abs(diff.fileDiff());
         int rankDiff = Math.abs(diff.rankDiff());
+        int fileDiff = Math.abs(diff.fileDiff());
 
-        if (fileDiff + rankDiff == 1) {
+        if (rankDiff == 2 && fileDiff == 1) {
             return true;
         }
 
-        return fileDiff == 1 && rankDiff == 1;
+        return rankDiff == 1 && fileDiff == 2;
     }
 }
