+++ src/test/java/chess/domain/piece/strategy/BishopLegalMoveCheckStrategyTest.java
@@ -1,7 +1,9 @@
-package chess.domain.strategy;
+package chess.domain.piece.strategy;
 
 import chess.domain.board.Board;
 import chess.domain.piece.Piece;
+import chess.domain.piece.strategy.BishopLegalMoveCheckStrategy;
+import chess.domain.piece.strategy.LegalMoveCheckStrategy;
 import chess.domain.square.File;
 import chess.domain.square.Rank;
 import chess.domain.square.Square;
@@ -11,39 +13,39 @@ import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-@DisplayName("나이트 움직임 전략")
-class KnightLegalMoveCheckStrategyTest {
+@DisplayName("비숍 움직임 전략")
+class BishopLegalMoveCheckStrategyTest {
 
     private LegalMoveCheckStrategy legalMoveCheckStrategy;
     private Board board;
 
     @BeforeEach
     void setUp() {
-        legalMoveCheckStrategy = new KnightLegalMoveCheckStrategy();
+        legalMoveCheckStrategy = new BishopLegalMoveCheckStrategy();
         board = new Board();
     }
 
-    @DisplayName("나이트는 올바른 위치가 들어오면 이동 가능을 반환한다.")
+    @DisplayName("비숍은 대각선 방향의 위치가 들어오면 이동 가능을 반환한다.")
     @Test
-    void knightCorrectDestination() {
+    void bishopCorrectDestination() {
         // given
-        Square source = Square.of(File.B, Rank.EIGHT);
-        Square destination = Square.of(File.C, Rank.SIX);
+        Square source = Square.of(File.E, Rank.SIX);
+        Square destination = Square.of(File.H, Rank.THREE);
         Piece destinationPiece = board.findPieceBySquare(destination);
 
         // when
-        boolean actual = legalMoveCheckStrategy.check(source, destination, board);
+        boolean actual = legalMoveCheckStrategy.check(source, destination,board);
 
         // then
         assertThat(actual).isTrue();
     }
 
-    @DisplayName("나이트는 잘못된 위치가 들어오면 이동 불가능을 반환한다.")
+    @DisplayName("비숍은 대각선 방향이 아닌 위치가 들어오면 이동 불가능을 반환한다.")
     @Test
-    void knightInCorrectDestination() {
+    void bishopInCorrectDestination() {
         // given
-        Square source = Square.of(File.B, Rank.EIGHT);
-        Square destination = Square.of(File.B, Rank.SIX);
+        Square source = Square.of(File.C, Rank.EIGHT);
+        Square destination = Square.of(File.C, Rank.THREE);
         Piece destinationPiece = board.findPieceBySquare(destination);
 
         // when
