+++ src/test/java/chess/domain/piece/strategy/PawnLegalMoveCheckStrategyTest.java
@@ -1,7 +1,9 @@
-package chess.domain.strategy;
+package chess.domain.piece.strategy;
 
 import chess.domain.board.Board;
 import chess.domain.piece.Piece;
+import chess.domain.piece.strategy.LegalMoveCheckStrategy;
+import chess.domain.piece.strategy.PawnLegalMoveCheckStrategy;
 import chess.domain.square.File;
 import chess.domain.square.Rank;
 import chess.domain.square.Square;
@@ -11,24 +13,24 @@ import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-@DisplayName("룩 움직임 전략")
-public class RookLegalMoveCheckStrategyTest {
+@DisplayName("폰 움직임 전략")
+public class PawnLegalMoveCheckStrategyTest {
 
     private LegalMoveCheckStrategy legalMoveCheckStrategy;
     private Board board;
 
     @BeforeEach
     void setUp() {
-        legalMoveCheckStrategy = new RookLegalMoveCheckStrategy();
+        legalMoveCheckStrategy = new PawnLegalMoveCheckStrategy();
         board = new Board();
     }
 
-    @DisplayName("룩은 수평 이동 위치가 들어오면 이동 가능을 반환한다.")
+    @DisplayName("폰은 올바른 위치가 입력되면 이동 가능을 반환한다.")
     @Test
-    void rookHorizontalDestination() {
+    void pawnCorrectDestination() {
         // given
-        Square source = Square.of(File.A, Rank.SIX);
-        Square destination = Square.of(File.H, Rank.SIX);
+        Square source = Square.of(File.A, Rank.THREE);
+        Square destination = Square.of(File.A, Rank.FOUR);
         Piece destinationPiece = board.findPieceBySquare(destination);
 
         // when
@@ -38,27 +40,12 @@ public class RookLegalMoveCheckStrategyTest {
         assertThat(actual).isTrue();
     }
 
-    @DisplayName("룩은 수직 이동 위치가 들어오면 이동 가능을 반환한다.")
+    @DisplayName("폰은 잘못된 위치가 입력되면 이동 불가능을 반환한다.")
     @Test
-    void rookVerticalDestination() {
+    void pawnInCorrectDestination() {
         // given
-        Square source = Square.of(File.A, Rank.FIVE);
-        Square destination = Square.of(File.A, Rank.TWO);
-        Piece destinationPiece = board.findPieceBySquare(destination);
-
-        // when
-        boolean actual = legalMoveCheckStrategy.check(source, destination, board);
-
-        // then
-        assertThat(actual).isTrue();
-    }
-
-    @DisplayName("룩은 잘못된 위치가 들어오면 이동 불가능을 반환한다.")
-    @Test
-    void rookInCorrectDestination() {
-        // given
-        Square source = Square.of(File.A, Rank.FIVE);
-        Square destination = Square.of(File.B, Rank.TWO);
+        Square source = Square.of(File.A, Rank.THREE);
+        Square destination = Square.of(File.C, Rank.FOUR);
         Piece destinationPiece = board.findPieceBySquare(destination);
 
         // when
@@ -66,5 +53,20 @@ public class RookLegalMoveCheckStrategyTest {
 
         // then
         assertThat(actual).isFalse();
+    }
+
+    @DisplayName("폰은 상대 말이 있을 경우 대각선 1칸이 입력되면 이동 가능을 반환한다.")
+    @Test
+    void moveOneDiagonal() {
+        // given
+        Square source = Square.of(File.A, Rank.THREE);
+        Square destination = Square.of(File.B, Rank.FOUR);
+        Piece destinationPiece = board.findPieceBySquare(destination);
+
+        // when
+        boolean actual = legalMoveCheckStrategy.check(source, destination, board);
+
+        // then
+        assertThat(actual).isTrue();
     }
 }
