+++ src/test/java/chess/domain/piece/strategy/RookLegalMoveCheckStrategyTest.java
@@ -1,7 +1,9 @@
-package chess.domain.strategy;
+package chess.domain.piece.strategy;
 
 import chess.domain.board.Board;
 import chess.domain.piece.Piece;
+import chess.domain.piece.strategy.LegalMoveCheckStrategy;
+import chess.domain.piece.strategy.RookLegalMoveCheckStrategy;
 import chess.domain.square.File;
 import chess.domain.square.Rank;
 import chess.domain.square.Square;
@@ -11,23 +13,23 @@ import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-@DisplayName("퀸 움직임 전략")
-class QueenLegalMoveCheckStrategyTest {
+@DisplayName("룩 움직임 전략")
+public class RookLegalMoveCheckStrategyTest {
 
     private LegalMoveCheckStrategy legalMoveCheckStrategy;
     private Board board;
 
     @BeforeEach
     void setUp() {
-        legalMoveCheckStrategy = new QueenLegalMoveCheckStrategy();
+        legalMoveCheckStrategy = new RookLegalMoveCheckStrategy();
         board = new Board();
     }
 
-    @DisplayName("퀸은 수평 이동이 들어오면 이동 가능을 반환한다.")
+    @DisplayName("룩은 수평 이동 위치가 들어오면 이동 가능을 반환한다.")
     @Test
-    void horizontalMove() {
+    void rookHorizontalDestination() {
         // given
-        Square source = Square.of(File.D, Rank.SIX);
+        Square source = Square.of(File.A, Rank.SIX);
         Square destination = Square.of(File.H, Rank.SIX);
         Piece destinationPiece = board.findPieceBySquare(destination);
 
@@ -38,12 +40,12 @@ class QueenLegalMoveCheckStrategyTest {
         assertThat(actual).isTrue();
     }
 
-    @DisplayName("퀸은 수직 이동이 들어오면 이동 가능을 반환한다.")
+    @DisplayName("룩은 수직 이동 위치가 들어오면 이동 가능을 반환한다.")
     @Test
-    void verticalMove() {
+    void rookVerticalDestination() {
         // given
-        Square source = Square.of(File.D, Rank.SIX);
-        Square destination = Square.of(File.D, Rank.THREE);
+        Square source = Square.of(File.A, Rank.FIVE);
+        Square destination = Square.of(File.A, Rank.TWO);
         Piece destinationPiece = board.findPieceBySquare(destination);
 
         // when
@@ -53,27 +55,12 @@ class QueenLegalMoveCheckStrategyTest {
         assertThat(actual).isTrue();
     }
 
-    @DisplayName("퀸은 대각 이동이 들어오면 이동 가능을 반환한다.")
+    @DisplayName("룩은 잘못된 위치가 들어오면 이동 불가능을 반환한다.")
     @Test
-    void diagonalMove() {
+    void rookInCorrectDestination() {
         // given
-        Square source = Square.of(File.D, Rank.SIX);
-        Square destination = Square.of(File.G, Rank.THREE);
-        Piece destinationPiece = board.findPieceBySquare(destination);
-
-        // when
-        boolean actual = legalMoveCheckStrategy.check(source, destination, board);
-
-        // then
-        assertThat(actual).isTrue();
-    }
-
-    @DisplayName("퀸은 잘못된 이동이 들어오면 이동 불가능을 반환한다.")
-    @Test
-    void cannotMove() {
-        // given
-        Square source = Square.of(File.D, Rank.EIGHT);
-        Square destination = Square.of(File.C, Rank.SIX);
+        Square source = Square.of(File.A, Rank.FIVE);
+        Square destination = Square.of(File.B, Rank.TWO);
         Piece destinationPiece = board.findPieceBySquare(destination);
 
         // when
